requirements.txt
google-cloud-secret-manager==2.16.0
hubspot-api-client==8.2.0
functions-framework==3.5.0
flask==3.0.0
requests==2.31.0
setuptools==69.0.0
